First Changes
I added a dev environment for testing purposes. I altered a bit of the apperance. To start the test environment open Terminal, go to directory with cd cmd and type "./start_both.sh". This will start the Dev view and the User view on a sepparate port. It shouldnt interfere with the adminer or any of the systems that are already in use.